tree: update kernel-muzel-14356307 a16 2025-12

+++ google-modules/wlan/bcmdhd/bcm4383/dhd_cfg80211.c
@@ -565,7 +565,7 @@ int dhd_cfgvendor_priv_string_handler(struct bcm_cfg80211 *cfg, struct wireless_
 	dhd_pub_t *dhd;
 	dhd_ioctl_t ioc = { 0, NULL, 0, 0, 0, 0, 0};
 	int ret = 0;
-	int8 index;
+	int index;
 
 	WL_TRACE(("entry: cmd = %d\n", nlioc->cmd));
